Major update to the API and implementation
Based on the newer code that got developed in cabal-install
and the new hackage-server. It has more compositional API
and the tar format compatability is better (but not perfect).
The documentation is also substantially better.
Tasks remaining: check full byte-for-byte compatability and
account for format quirks in other tools. Check for security,
tarbombs and portability. Use tar torture test in a testsuite.